requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.3.0
tensorflow>=2.12.0
shap>=0.41.0
streamlit>=1.22.0
schedule>=1.2.0
joblib>=1.2.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pyarrow>=10.0.0
python-dotenv>=1.0.0
streamlit-folium>=0.13.0
folium>=0.14.0
geopy>=2.3.0
flask>=2.3.0
flask-cors>=4.0.0